Crate executor

Structs

CommittableBlock

ExecutedState
Executed state derived from StateComputeResult that is maintained with every proposed block. state_id(transaction accumulator root hash) summarized both the information of the version and the validators.

ExecutedTrees

Executor
Executor implements all functionalities the execution module needs to provide.

ProcessedVMOutput
Generated by processing VM's output.

StateComputeResult
A structure that summarizes the result of the execution needed for consensus to agree on. The execution is responsible for generating the ID of the new state, which is returned in the result.

TransactionData
The entire set of data associated with a transaction. In addition to the output generated by VM which includes the write set and events, this also has the in-memory trees.