Trait executor::BatchExecutionState

pub trait BatchExecutionState: ExecutionState {
    fn load_next_certificate_index<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<SequenceNumber, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_consensus<'life0, 'life1, 'async_trait>(
        &'life0 self,
        consensus_output: &'life1 ConsensusOutput,
        transaction_batches: Vec<Vec<SerializedTransaction>>
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

Execution state that gets whole certificates and the corresponding batches for execution. It is responsible for deduplication in case the same certificate is re-delivered after a crash.

Required Methods

fn load_next_certificate_index<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<SequenceNumber, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load the last consensus index from storage.

It should return the index from which it expects a replay, so one higher than the last certificate index it successfully committed. This is so it has the same semantics as ExecutionIndices.

fn handle_consensus<'life0, 'life1, 'async_trait>(
&'life0 self,
consensus_output: &'life1 ConsensusOutput,
transaction_batches: Vec<Vec<SerializedTransaction>>
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Execute the transactions and atomically persist the consensus index.

TODO: This function should be allowed to return a new committee to reconfigure the system.