Trait executor::ExecutionState

pub trait ExecutionState {
    type Error: ExecutionStateError;

    fn ask_consensus_write_lock(&self) -> bool;
    fn release_consensus_write_lock(&self);
}

Required Associated Types

type Error: ExecutionStateError

The error type to return in case something went wrong during execution.

Required Methods

fn ask_consensus_write_lock(&self) -> bool

Simple guardrail ensuring there is a single instance using the state to call handle_consensus_transaction. Many instances may read the state, or use it for other purposes.

fn release_consensus_write_lock(&self)

Tell the state that the caller instance is no longer using calling