Crate executor

Macros

bail

ensure

try_fut_and_permit

Structs

ExecutionIndices

The state of the subscriber keeping track of the transactions that have already been executed. It ensures we do not process twice the same transaction despite crash-recovery.

Executor

A client subscribing to the consensus output and executing every transaction.

Enums

SubscriberError

Traits

ExecutionState

ExecutionStateError

Trait do separate execution errors in two categories: (i) errors caused by a bad client, (ii) errors caused by a fault in the authority.

Functions

get_restored_consensus_output

Type Definitions

ExecutorOutput

The output of the executor.

SerializedTransaction

Convenience type representing a serialized transaction.

SerializedTransactionDigest

Convenience type representing a serialized transaction digest.

SubscriberResult