executor

Function process_write_set

pub fn process_write_set(
    transaction: &Transaction,
    account_to_state: &mut HashMap<AccountAddress, AccountState>,
    write_set: WriteSet,
) -> Result<HashMap<AccountAddress, AccountStateBlob>>

For all accounts modified by this transaction, find the previous blob and update it based on the write set. Returns the blob value of all these accounts.