Trait executor::ExecutionState

pub trait ExecutionState {
    type Transaction: DeserializeOwned + Send + Debug;
    type Error: ExecutionStateError;
    type Outcome;

    fn handle_consensus_transaction<'life0, 'life1, 'async_trait>(
        &'life0 self,
        consensus_output: &'life1 ConsensusOutput,
        execution_indices: ExecutionIndices,
        transaction: Self::Transaction
    ) -> Pin<Box<dyn Future<Output = Result<Self::Outcome, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn deserialize(bytes: &[u8]) -> Result<Self::Transaction, Error>;
    fn ask_consensus_write_lock(&self) -> bool;
    fn release_consensus_write_lock(&self);
    fn load_execution_indices<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<ExecutionIndices, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required Associated Types

type Transaction: DeserializeOwned + Send + Debug

The type of the transaction to process.

type Error: ExecutionStateError

The error type to return in case something went wrong during execution.

type Outcome

The execution outcome to output.

Required Methods

fn handle_consensus_transaction<'life0, 'life1, 'async_trait>(
&'life0 self,
consensus_output: &'life1 ConsensusOutput,
execution_indices: ExecutionIndices,
transaction: Self::Transaction
) -> Pin<Box<dyn Future<Output = Result<Self::Outcome, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Execute the transaction and atomically persist the consensus index. This function returns an execution outcome that will be output by the executor channel. It may also return a new committee to reconfigure the system.

fn deserialize(bytes: &[u8]) -> Result<Self::Transaction, Error>

Deserialize the message bytes into Transaction type. This allows an implementation of ExecutionState to customize how to deserialize the message, in case customized serialization was used when sending the message.

fn ask_consensus_write_lock(&self) -> bool

Simple guardrail ensuring there is a single instance using the state to call handle_consensus_transaction. Many instances may read the state, or use it for other purposes.

fn release_consensus_write_lock(&self)

Tell the state that the caller instance is no longer using calling

fn load_execution_indices<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<ExecutionIndices, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Load the last consensus index from storage.